# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PRC API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "api/"
        }
    ],
    "paths": {
        "/evaluator-status-change": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "evaluator Status Change",
                "description": "evaluator Status Change",
                "operationId": "evaluatorStatusChange",
                "requestBody": {
                    "description": "Status and Evaluator Id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "evaluator_id",
                                    "status"
                                ],
                                "properties": {
                                    "evaluator_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password reset successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "apiKey",
                "description": "Authentication Token",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}